requirements.txt
Django==4.2.1
django-crispy-forms==1.14.0
pytesseract==0.3.10
opencv-python==4.6.0.66
imutils==0.5.4
python-dateutil==2.8.2
Pillow==9.3.0
numpy==1.23.5
psycopg2-binary==2.9.5
scipy==1.9.3
python-dotenv==1.0.0
six==1.16.0
xlwt==1.3.0
python-dateutil==2.8.2
validate_email==1.3
psycopg2==2.9.6
python-dotenv==1.0.0